+++ CHANGELOG.rst
@@ -11,6 +11,11 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 `Unreleased`_
 -------------
 
+Added
+~~~~~~~
+
+- Added ``orientation`` parameter to ``RealSphericalHarmonicsDirectivity`` object.
+
 Fixed
 ~~~~~